+++ cmd/wire.go
@@ -3,61 +3,65 @@
 package main
 
 import (
-	"os"
-	"strconv"
+  "os"
+  "strconv"
 
-	"github.com/google/wire"
-	"github.com/kalilventura/vehicle-management-payment/internal/payments"
-	"github.com/kalilventura/vehicle-management-payment/internal/shared/domain/entities"
-	"github.com/kalilventura/vehicle-management-payment/internal/shared/infrastructure/configuration"
+  "github.com/google/wire"
+  "github.com/kalilventura/vehicle-management-payment/internal/payments"
+  "github.com/kalilventura/vehicle-management-payment/internal/shared/domain/entities"
+  "github.com/kalilventura/vehicle-management-payment/internal/shared/infrastructure/configuration"
+  "github.com/kalilventura/vehicle-management-payment/internal/shared/infrastructure/services"
 )
 
-func InjectModules() []entities.HTTPModule {
-	wire.Build(
-		injectWebhookSettings,
-		injectPaymentSettings,
-		injectDatabaseSettings,
-		configuration.NewDatabaseClient,
-		payments.Container,
-		newModules,
-	)
-	return nil
+func InjectApp() *App {
+  wire.Build(
+    InjectSettings,
+    injectWebhookSettings,
+    injectPaymentSettings,
+    injectDatabaseSettings,
+    configuration.NewDatabaseClient,
+    services.Container,
+    payments.Container,
+    newModules,
+    NewApp,
+  )
+  return nil
 }
 
 func InjectSettings() *entities.Settings {
-	port, _ := strconv.Atoi(os.Getenv("PORT"))
-	return &entities.Settings{port}
+  port, _ := strconv.Atoi(os.Getenv("PORT"))
+  return &entities.Settings{Port: port}
 }
 
 func injectDatabaseSettings() *entities.DatabaseSettings {
-	host := os.Getenv("DB_HOST")
-	name := os.Getenv("DB_NAME")
-	port := os.Getenv("DB_PORT")
-	user := os.Getenv("DB_USER")
-	password := os.Getenv("DB_PASSWORD")
-	dbSSL := os.Getenv("DB_SSL")
-	return entities.NewDatabaseSettings(
-		host,
-		name,
-		port,
-		user,
-		password,
-		dbSSL,
-	)
+  host := os.Getenv("DB_HOST")
+  name := os.Getenv("DB_NAME")
+  port := os.Getenv("DB_PORT")
+  user := os.Getenv("DB_USER")
+  password := os.Getenv("DB_PASSWORD")
+  dbSSL := os.Getenv("DB_SSL")
+  return entities.NewDatabaseSettings(
+    host,
+    name,
+    port,
+    user,
+    password,
+    dbSSL,
+  )
 }
 
 func injectPaymentSettings() *entities.PaymentSettings {
-	stripeKey := os.Getenv("STRIPE_KEY")
-	return &entities.PaymentSettings{stripeKey}
+  stripeKey := os.Getenv("STRIPE_KEY")
+  return &entities.PaymentSettings{StripeKey: stripeKey}
 }
 
 func injectWebhookSettings() *entities.WebhookSettings {
-	stripeKey := os.Getenv("STRIPE_WEBHOOK_KEY")
-	return &entities.WebhookSettings{stripeKey}
+  stripeKey := os.Getenv("STRIPE_WEBHOOK_KEY")
+  return &entities.WebhookSettings{Secret: stripeKey}
 }
 
 func newModules(paymentsModule *payments.Module) []entities.HTTPModule {
-	return []entities.HTTPModule{
-		paymentsModule,
-	}
+  return []entities.HTTPModule{
+    paymentsModule,
+  }
 }